Compute lat and lon 2d coordinate arrays on range azimuth grids.

+++ met/src/tools/tc_utils/tc_rmw/tc_rmw.h
@@ -114,8 +114,8 @@ static Met2dDataFileFactory mtddf_factory;
 static Met2dDataFile* fcst_mtddf = (Met2dDataFile*) 0;
 
 // Grid coordinate arrays
-static float* lat_grid;
-static float* lon_grid;
+static double* lat_grid;
+static double* lon_grid;
 
 ////////////////////////////////////////////////////////////////////////
